+++ src/game/units/wizard/archetypes/arcanorouter/mod.rs
@@ -8,4 +8,5 @@ mod systems;
 pub use components::ArcanoRouterBonuses;
 pub use messages::SliderAdjustMessage;
 pub(in crate::game) use plugin::ArcanoRouterPlugin;
+pub(in crate::game) use resources::ArcanoRouterSetupBaseline;
 pub use resources::{ArcanoRouterState, SliderType};